+++ src/model/AccountsProfile/types.ts
@@ -1,8 +1,8 @@
-import {components} from 'src/types/model'
+import {components} from '../../types/model'
 
-export interface AccountsProfileGetParams {
-    id: string,
-    [index: string]: any
-  };
 export type ProfileType = components['schemas']['Profile']
 
+export interface AccountsProfileGetParams {
+  id: string,
+  [index: string]: any
+};
