+++ src/model/git/Blob/index.ts
@@ -1,0 +1,4 @@
+import Blob from './Blob'
+export * from './types'
+
+export default Blob
