+++ src/model/git/Blob/types.ts
@@ -1,0 +1,8 @@
+import {components} from 'src/types/model'
+
+export type BlobType = components['schemas']['Blob']
+
+export type BlobParams = {
+    projectId: string,
+    sha: string
+  };
