+++ src/model/git/Commit/index.ts
@@ -1,0 +1,4 @@
+import Commit from "./Commit";
+export * from './types'
+
+export default Commit
