+++ src/model/git/Commit/types.ts
@@ -1,0 +1,8 @@
+import {components} from 'src/types/model'
+
+export type CommitType = components['schemas']['Commit']
+
+export interface CommitParams {
+    projectId?: string,
+    [index: string]: any
+  }
