+++ src/model/git/Tree/index.ts
@@ -1,0 +1,6 @@
+import Tree from './Tree'
+
+export default Tree
+export * from './types'
+
+
