+++ src/model/git/Tree/types.ts
@@ -1,0 +1,8 @@
+import {components} from '../../../types/model'
+
+export type TreeType = components['schemas']['Tree']
+
+export type TreeParams = {
+    projectId?: string,
+    [key: string]: any
+  };
